quic: dedicate the data-types.h file to quic
