+++ CMakeLists.txt
@@ -33,7 +33,7 @@ SET(SOURCE_FILES
   "http.cc"
   "json.h"
   "json.cc"
-  "data-types.h"
+  "quic.h"
 )
 
 ADD_CUSTOM_COMMAND(
